refactor: 코드가독성 향상을 위한 코드 리팩토링 진행
- 클래스, 메서드 및 변수명 변경
- delete: 필요없는 클래스파일 삭제
- 예외 메시지 ENUM 클래스에서 중앙 관리

+++ src/main/java/lotto/dto/MatchInfo.java
@@ -1,7 +1,7 @@
 package lotto.dto;
 
 public record MatchInfo(
-        int matchCount,
+        long matchCount,
         boolean isMatchBonusNumber
 ) {
